requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
lxml==4.9.3
multidict==6.0.4
numpy==1.25.1
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
soupsieve==2.4.1
tzdata==2023.3
yarl==1.9.2
